dpo_python_django-master/M_17_import_export/requirements.txt
apt-xapian-index==0.49
asgiref==3.6.0
attrs==23.1.0
blinker==1.4
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.2
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.2
flake8==6.0.0
flake8-docstrings==1.7.0
fuse-python==1.0.2
gpg===1.16.0-unknown
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydocstyle==6.3.0
pyflakes==3.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.19.3
pyTelegramBotAPI==4.11.0
python-apt==2.4.0+ubuntu1
python-debian===0.1.43ubuntu1
pytz==2023.3
pyxattr==0.7.2
PyYAML==5.4.1
reportlab==3.6.8
requests==2.30.0
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
systemd-python==234
tornado==6.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.2
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
